# api/swagger.yml
openapi: '3.0.3'
info:
  title: Breeds API
  version: '1.0'
servers:
  - url: http://localhost:50010/v1
tags:
  - name: Breeds
    description: operations on breeds resource
paths:
  /breeds:
    get:
      tags:
        - Breeds
      summary: List breeds
      description: Perform a query on breeds base on pet characteristic (weight and species)
      operationId: ListBreeds
      parameters:
        - $ref: "#/components/parameters/Species"
        - $ref: "#/components/parameters/AverageFemaleAdultWeight"
        - $ref: "#/components/parameters/AverageMaleAdultWeight"
        - $ref: "#/components/parameters/PetSize"
      responses:
        '200':
          $ref: "#/components/responses/BreedsList"
        '500':
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - Breeds
      summary: Create one breed
      description: Create one breed if it not exists. Othewise it return an error
      operationId: CreateOneBreed
      requestBody:
        $ref: "#/components/requestBodies/Breed"
      responses:
        '201':
          $ref: "#/components/responses/BreedResponse"
        '400':
          $ref: "#/components/responses/BadRequestError"
        '409':
          $ref: "#/components/responses/ResourceAlreadyExistsError"
        '500':
          $ref: "#/components/responses/InternalServerError"
      
  /breeds/name/{breed_name}:
    get:
      tags:
        - Breeds
      summary: Retrieve a given breed by its name
      description: Retrieve a given breed by its name if this resource exists
      operationId: GetBreedByName
      parameters:
      - $ref: "#/components/parameters/BreedName"
      responses:
        '200':
          $ref: "#/components/responses/BreedResponse"
        '400':
          $ref: "#/components/responses/BadRequestError"
        '404':
          $ref: "#/components/responses/ResourceNotFoundError"
        '500':
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - Breeds
      summary: Delete a given breed by its name
      description: Delete a given breed by its name if this resource exists
      operationId: DeleteBreedByName
      parameters:
      - $ref: "#/components/parameters/BreedName"
      responses:
        '200':
          $ref: "#/components/responses/BreedResponse"
        '400':
          $ref: "#/components/responses/BadRequestError"
        '404':
          $ref: "#/components/responses/ResourceNotFoundError"
        '500':
          $ref: "#/components/responses/InternalServerError"
    put:
      tags:
        - Breeds
      summary: Update or create one breed
      description: If the breed exists it update this breed with the entire body of the request. Othewise it creates it.
      operationId: CreateOrUpdateBreedByName
      parameters:
      - $ref: "#/components/parameters/BreedName"
      requestBody:
        $ref: "#/components/requestBodies/Breed"
      responses:
        '200':
          description: Update successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Breeds"
        '201':
          description: Breed created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Breeds"
        '204':
          description: Breed exists but no updates has been done
        '400':
          $ref: "#/components/responses/BadRequestError"
        '404':
          $ref: "#/components/responses/ResourceNotFoundError"
        '500':
          $ref: "#/components/responses/InternalServerError"
        
components:
  requestBodies:
    Breed:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Breeds"
  parameters:
    PetSize:
      in: query
      required: false
      name: "pet_size"
      schema:
        $ref: "#/components/schemas/PetSize"
    AverageMaleAdultWeight:
      in: query
      required: false
      name: "average_male_adult_weight"
      schema:
        type: integer
        description: Average weight of the male adult in gramme
        example: 2000
        minimum: 0
    AverageFemaleAdultWeight:
      in: query
      required: false
      name: "average_female_adult_weight"
      schema:
        type: integer
        description: Average weight of the female adult in gramme
        example: 1000
        minimum: 0
    Species:
      in: query
      name: species
      required: false
      schema:
        $ref: "#/components/schemas/Species"
    BreedName:
      in: path
      required: true
      name: breed_name
      schema:
        type: string
        pattern: "[a-z]+(_[a-z]+)*"
        minLength: 2
        maxLength: 255
  responses:
    BreedsList:
      description: Response when the request is successful
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "#/components/schemas/Breeds"
    BreedResponse:
      description: Response when the request is successful
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Breeds"
    ResourceAlreadyExistsError:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ResourceNotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequestError:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  schemas:
    Species:
      type: string
      enum:
        - dog
        - cat
    PetSize:
      type: string
      enum:
        - small
        - medium
        - tall
      description: size of the pet
    Error:
      type: object
      additionalProperties: false
      required:
        - message
      properties:
        message:
          type: string
          minLength: 2
          example: "error message"
    Breeds:
      type: object
      additionalProperties: false
      required:
        - species
        - pet_size
        - name
      properties:
        pet_size:
          $ref: "#/components/schemas/PetSize"
        species:
          $ref: "#/components/schemas/Species"
        name:
          type: string
          description: Name of the breed. Should be in snake case and should be unique
          pattern: "[a-z]+(_[a-z]+)*"
          minLength: 2
          maxLength: 255
          example: "polish_hunting_dog_kopov"
        average_male_adult_weight:
          type: integer
          description: Average weight of the male adult in gramme
          example: 2000
          minimum: 0
          default: 0
        average_female_adult_weight:
          type: integer
          description: Average weight of the female adult in gramme
          example: 1000
          minimum: 0
          default: 0